refactor(ethers-providers): colocate platform seams with the types they serve
Replaces the catch-all PlatformApi files with one declaration per type. Each
expect now sits directly above the type that extends it - PlatformRpcRequest in
RpcRequest.kt, PlatformRpcSubscribe in RpcSubscribe.kt, PlatformPendingInclusion
in PendingInclusion.kt - so the seam is visible where it matters instead of in a
separate file.

The jvmSharedMain actuals are split to match, one file each. They are named
after the class rather than the type they serve, because ktlint requires a file
holding a single class to be named after it.
